Update lightCustom.scss - moved .active into div.sidebar
The .active >*>* was causing issues with tab sets since the active tab would then style according to this. By moving it into the div.sidebar css now it will only apply to that particular place.  Also, the color changed to #314D64 to make it more accessible. I also made the font wt to 'bolder'