Update darkCustom.scss  moved .active into the div.sidebar css
The .active >> was causing issues with tab sets since the active tab would then style according to this. By moving it into the div.sidebar css now it will only apply to that particular place. Also, the color changed to white. I also made the font wt to 'bolder' .active > * > *{